+++ packages/harpguru-core/src/components/menu-of-tunings/utils/reduce-for-new-harp-strata-by-tuning/index.ts
@@ -1,0 +1,1 @@
+export { reduceForNewHarpStrataByTuning } from './reduce-for-new-harp-strata-by-tuning'