+++ packages/harpguru-core/src/components/menu-of-tunings/utils/reduce-for-new-harp-strata-by-valving/index.ts
@@ -1,0 +1,1 @@
+export { reduceForNewHarpStrataByValving } from './reduce-for-new-harp-strata-by-valving'