+++ packages/harpguru-core/CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to ~~[Semantic Versioning](https://semver.org/spec/v2.0
 
 ### Added
 
-- MINOR: Ability to display just a portion of the harp in the harpface (no UI control yet)
+- MINOR: Ability to zoom to set number of holes
 
 ### Changed